pyproject.toml
[project]
name = "sncf-mcp"
version = "0.1.0"
description = "MCP server for SNCF"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "requests>=2.32.0",
    "python-dateutil>=2.8.2",
    "selenium>=4.0.0",
    "undetected-chromedriver>=3.5.0",
    "setuptools>=65.0.0",  # Provides distutils for Python 3.12+
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-httpx>=0.32.0",
]
